Changed Icon. Completed Main Activity with dialog. Changed price to float type in GroceryItems.kt.

+++ app/src/main/java/com/kanaishk/groceryapp/GroceryItems.kt
@@ -14,7 +14,7 @@ data class GroceryItems (
     var itemQuantity: Int,
 
     @ColumnInfo(name = "itemPrice")
-    var itemPrice: Int,
+    var itemPrice: Float,
 
 ) {
     @PrimaryKey(autoGenerate = true)
